requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
packaging==24.2
pillow==11.0.0
pytesseract>=0.3.10
matplotlib==3.8.2
fastapi==0.109.1
python-multipart==0.0.6
uvicorn==0.27.0
jinja2==3.1.3
